Examples: 修正数据库示例，增加Int64的支持。

+++ Examples/Database/MailCSharp/Src/IDataAccess.cs
@@ -7,6 +7,6 @@ namespace Database.Database
 {
     public partial interface IDataAccess
     {
-        List<String> FromMailAttachmentSelectManyForNameById(int Id);
+        List<String> FromMailAttachmentSelectManyForNameById(Int64 Id);
     }
 }
